feat: Refactor tag filters to use a shared FilterList component for improved reusability
